expression float

+++ java-solutions/expression/Divide.java
@@ -12,7 +12,7 @@ public class Divide extends Operation {
     }
 
     @Override
-    public double operation(double a, double b) {
+    public float operation(float a, float b) {
         return a / b;
     }
 }
